rename props to data-model
